+++ controllers/teacherManagerControllers.js
@@ -2,48 +2,11 @@ const { ObjectId } = require("mongodb")
 const { getDb } = require("../utils/dbConnects")
 
 
-
-
-const getSingleTeacher = async(req,res) =>{
-    try{
-        const db = getDb()
-        const id = req.params.id
-        
-        if(!id){
-            return res.status(404).json({
-                message: 'Id required!' 
-            })
-        }
-
-        const query = { _id :new ObjectId(id) }
-        
-        const teacher = await db.collection('teacher').findOne(query)
-        
-        if(!teacher){
-            return res.status(404).json({
-                teacher,
-                 message: 'Invalid student id' 
-            })
-        }
-
-        res.status(200).json({
-            teacher,
-            message: 'successfull'
-        })
-
-    }catch(err){
-        console.log(err)
-    }
-}
-
-
 const getAllTeacher = async (req,res) =>{
     try{
         const db = getDb()
-        
         const teachers = await db.collection('teacher').find().toArray()
-        
-        // console.log(teachers)
+
         if(!teachers.length){
             return res.status(400).json({
                 teachers,
@@ -55,9 +18,7 @@ const getAllTeacher = async (req,res) =>{
             teachers,
             message: 'successfull!'
         })
-
-    }
-    catch(err){
+    }catch(err){
         console.log(err)
     }
 }
@@ -69,6 +30,14 @@ const EditTeacherData = async (req,res) =>{
         const { name , number , img_link } = req.body
         const id = req.params.id
 
+        const email_d = req.email 
+        const role_d  = req.role
+
+        if(!email_d && (role_d !== 'admin' || role_d !== 'editor')){
+            return res.status(404).json({
+                message : 'Unauthorized',
+            })
+        }
 
         const query = { _id :new ObjectId(id) }
 
@@ -80,9 +49,7 @@ const EditTeacherData = async (req,res) =>{
             })
         }
 
-
         const options = { upsert: true }
-
         const updateDoc = {
             $set: {
                 number : number ? number : teacher.number,
@@ -94,9 +61,7 @@ const EditTeacherData = async (req,res) =>{
         const result = await db.collection('teacher').updateMany(query, updateDoc, options)
 
         res.status(200).json(result)
-
-    }
-    catch(err){
+    }catch(err){
         console.log(err)
     }
 }
@@ -107,20 +72,26 @@ const insertTeachersData = async (req,res) =>{
         const db = getDb()
         const teachers_data = req.body
 
-
         if(!teachers_data){
             return res.status(404).json({
                 message : 'No student found',
             })
         }
 
-        const dataArray = Array.isArray(teachers_data) ? teachers_data : [teachers_data]
+        const email_d = req.email 
+        const role_d  = req.role
 
+        if(!email_d && (role_d !== 'admin' || role_d !== 'editor')){
+            return res.status(404).json({
+                message : 'Unauthorized',
+            })
+        }
+
+        const dataArray = Array.isArray(teachers_data) ? teachers_data : [teachers_data]
         const response = await db.collection('teacher').insertMany(dataArray)
 
         res.status(200).json(response)
-    }
-    catch(err){
+    }catch(err){
         console.log(err)
     }
 }
@@ -137,19 +108,24 @@ const deleteTeacher = async (req,res) =>{
             })
         }
 
+        const email_d = req.email 
+        const role_d  = req.role
+
+        if(!email_d && (role_d !== 'admin' || role_d !== 'editor')){
+            return res.status(404).json({
+                message : 'Unauthorized',
+            })
+        }
+
         const query = { _id :new ObjectId(id) }
-
         const delete_response = await db.collection('teacher').deleteOne(query)
-
         res.status(200).json(delete_response)
-    }
-    catch(err){
+    }catch(err){
         console.log(err)
     }
 }
 
 module.exports = {
-    getSingleTeacher,
     getAllTeacher,
     EditTeacherData,
     insertTeachersData,
